* src/java/org/archive/util/TmpDirTestCase.java     Test for null list of files.  Otherwise was getting NPE.

+++ trunk/heritrix/src/java/org/archive/util/TmpDirTestCase.java
@@ -131,9 +131,12 @@ public class TmpDirTestCase
     public void cleanUpOldFiles(File basedir, String basename)
     {
         File [] files = getListOfFiles(basedir, basename);
-        for (int i = 0; i < files.length; i++)
+        if (files != null)
         {
-            files[i].delete();
+            for (int i = 0; i < files.length; i++)
+            {
+                files[i].delete();
+            }
         }
     }
